requirements.txt
asgiref==3.8.1
Django==4.2.16
gunicorn==23.0.0
packaging==24.2
psycopg2-binary==2.9.9
sqlparse==0.5.2
typing_extensions==4.12.2
django-cockroachdb>=4.2,<4.3
dj-database-url==2.3.0
boto3==1.35.71